requirements.txt
azure-devops==6.0.0b4
azure-functions==1.11.2
certifi==2022.6.15.2
charset-normalizer==2.1.1
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.12
